+++ src/main/java/me/maxiiiiii/skyblockdragons/mining/material/types/MiningBlockMaterial.java
@@ -9,7 +9,7 @@ import me.maxiiiiii.skyblockdragons.player.skill.SkillType;
 import java.util.Arrays;
 
 public abstract class MiningBlockMaterial extends BlockMaterial {
-    public MiningBlockMaterial(String itemID, BlockMaterials material, int blockStrength, int breakingPower, double skillExp, int experience, BlockDrop... drops) {
+    public MiningBlockMaterial(String itemID, BlockMaterials material, double blockStrength, int breakingPower, double skillExp, int experience, BlockDrop... drops) {
         super(itemID, material, blockStrength, breakingPower, Arrays.asList(ItemType.PICKAXE, ItemType.DRILL), skillExp, SkillType.MINING, experience, drops);
     }
 }
